sfincs refactor
