+++ geb/reasonable_default_config.yml
@@ -18,6 +18,7 @@ general:
 hazards:
   floods:
     simulate: false
+    forcing_method: headwater_points  # headwater_points or preciptation
     resolution: 150
     nr_subgrid_pixels: 4
     crs: auto
